requirements.txt
python-telegram-bot==20.0
requests==2.32.0
flask==3.0.0